+++ exodus/urls.py
@@ -25,6 +25,7 @@ urlpatterns = [
     path("", include("news.urls")),
     path("", include("starmap.urls")),
     path("", include("starships.urls")),
+    path("", include("spacebattle.urls")),
     path("", include("comms.urls")),
     path("", include("characters.urls")),
     path("", include("agencies.urls")),